requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
google-genai
pypdf
requests
python-multipart
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pymupdf
